move PW() from main.h -> handle.h
