+++ lock/main.cpp
@@ -2,11 +2,13 @@
 
 int main()
 {
-    PW PW1;
-    led_green = !led_green;
-    led_red = !led_red;
     ticker_scan_column.attach(ScanColumn, SCAN_COLUMN_PERIOD);
-    PW1.ResetUserInput();
+
+    char user_password_administrator_default[MAX_PW_LENGTH] = {'1', '2', '3', '4', '5', '6', '7', '8', '9', '9'};
+    char user_password_default[MAX_PW_LENGTH] = {'5', '4', '3', '2', '1'};
+    PW PW1 (user_password_administrator_default, user_password_default);
+
+    PW1.ResetUserInputBuffer();
     DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
     while (true)
     {
@@ -30,17 +32,17 @@ int main()
                         PW1.GetUserInputBuffer();
                         PW1.UpdatePassword();
                         DisplayString("DONE");
-                        PW1.ResetUserInput();
-                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                         ticker_led.detach();
+                        PW1.ResetUserInputBuffer();
+                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                     }
                     else
                     {
                         ticker_led.attach(ToggleRedLED, 100ms);
                         DisplayString("WRONG");
-                        PW1.ResetUserInput();
-                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                         ticker_led.detach();
+                        PW1.ResetUserInputBuffer();
+                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                     }
                 }
                 else if (PW1.user_check_counter <= MAX_PW_INPUT_TIMES)
@@ -49,18 +51,18 @@ int main()
                     {
                         ticker_led.attach(ToggleGreenLED, 100ms);
                         DisplayString("PASS");
-                        PW1.ResetUserInput();
-                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                         ticker_led.detach();
+                        PW1.ResetUserInputBuffer();
+                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                         // #TODO 需要添加开门的函数
                     }
                     else
                     {
                         ticker_led.attach(ToggleRedLED, 100ms);
                         DisplayString("FAILED");
-                        PW1.ResetUserInput();
-                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                         ticker_led.detach();
+                        PW1.ResetUserInputBuffer();
+                        DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                     }
                 }
                 else
@@ -68,10 +70,10 @@ int main()
                     ticker_led.attach(ToggleRedLED, 100ms);
                     DisplayString("Please try again later");
                     ThisThread::sleep_for(PW_FALSE_SELLP_PERIOD);
-                    PW1.ResetUserInput();
+                    ticker_led.detach();
+                    PW1.ResetUserInputBuffer();
                     DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
                     PW1.user_check_counter = 0;
-                    ticker_led.detach();
                 }
                 // display.clear();
                 // display.Home();
@@ -85,7 +87,7 @@ int main()
             }
             else if (PW1.user_input == '*')
             {
-                PW1.ResetUserInput();
+                PW1.ResetUserInputBuffer();
                 DisplayInput(PW1.user_input_buffer, PW1.user_input_counter);
             }
             else
